Removed the old trendline component and completed all the final changes
