+++ packages/esm-patient-test-results-app/src/trendline/trendline.scss
@@ -17,7 +17,7 @@
   display: flex;
   justify-content: center;
   align-items: baseline;
-  width: calc(100% - 22rem);
+  flex-grow: 1;
 }
 
 .title {
@@ -32,21 +32,26 @@
 }
 
 :global(.bx--cc--area path.area) {
-  stroke: none !important;
-  fill: #a7f0ba !important;
+  stroke: none;
+  fill: #a7f0ba;
 }
 
 .header {
-  display: flex;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
   align-items: center;
   background: white;
   border-bottom: 1px solid $ui-03;
 }
 
-.back-button {
+.backButton {
   align-items: center;
   display: flex;
   justify-content: flex-start;
+
+  a {
+    text-decoration: none;
+  }
 
   button {
     display: flex;
@@ -62,18 +67,17 @@
   }
 }
 
-.Background {
-  width: 100%;
-  // height: 20.125rem;
-  // margin: 3rem 12rem 1rem;
-  padding: 2.375rem 5.688rem 2.188rem 1rem;
-  background: white;
+.inlineLoader {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  padding-right: $spacing-05;
 }
 
-.Actions-Navigation-arrow--left-16 {
-  width: 1.25rem;
-  height: 1.25rem;
-  background-color: var(--open-mrs-background-gray);
+.background {
+  width: 100%;
+  padding: 2.375rem 5.688rem 2.188rem 1rem;
+  background: white;
 }
 
 .range-tabs :global(.bx--tabs--scrollable__nav-link) {
